requirements.txt
fastapi==0.116.1
google-generativeai==0.8.5
numpy==2.2.6
opencv-python==4.12.0.88
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.4
ultralytics==8.3.165
uvicorn==0.35.0